+++ UbiClubFeedbackApp/UbiClub.Feedback.Tests/ModelFactory/FeedbackCreateModelFactoryTests.cs
@@ -6,7 +6,7 @@ using UbiClub.Feedback.Api.Interfaces;
 using UbiClub.Feedback.Api.ModelFactory;
 using UbiClub.Feedback.Tests.Helpers;
 
-namespace UbiClub.Feedback.Tests.ModelFactories
+namespace UbiClub.Feedback.Tests.ModelFactory
 {
     [TestFixture]
     internal class FeedbackCreateModelFactoryTests : BaseTest
